Add database as dependency in Run() function
Use DI to make use of best practices and to enable unit testing.
